refactor: move all of connect.rs to editor.rs to keep things more together
