--- daemon/src/connect.rs
@@ -1,43 +1,0 @@
-use crate::daemon::DocumentActorHandle;
-use crate::editor::spawn_editor_connection;
-use crate::sandbox;
-use std::io;
-use std::path::{Path, PathBuf};
-use tokio::net::UnixListener;
-use tracing::info;
-
-/// # Panics
-///
-/// Will panic if we fail to listen on the socket, or if we fail to accept an incoming connection.
-pub async fn make_editor_connection(socket_path: PathBuf, document_handle: DocumentActorHandle) {
-    // Using the sandbox method here is technically unnecessary,
-    // but we want to really run all path operations through the sandbox module.
-    if sandbox::exists(Path::new("/"), Path::new(&socket_path))
-        .expect("Failed to check existence of path")
-    {
-        sandbox::remove_file(Path::new("/"), &socket_path).expect("Could not remove socket");
-    }
-    let result = accept_editor_loop(&socket_path, document_handle).await;
-    match result {
-        Ok(()) => {}
-        Err(err) => {
-            panic!("Failed to make editor connection: {err}");
-        }
-    }
-}
-
-async fn accept_editor_loop(
-    socket_path: &Path,
-    document_handle: DocumentActorHandle,
-) -> Result<(), io::Error> {
-    let listener = UnixListener::bind(socket_path)?;
-    info!("Listening on UNIX socket: {}", socket_path.display());
-
-    loop {
-        let (stream, _addr) = listener.accept().await?;
-
-        let id = document_handle.next_editor_id();
-
-        spawn_editor_connection(stream, document_handle.clone(), id).await;
-    }
-}
