+++ daemon/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod actors;
-pub mod connect;
 pub mod daemon;
 pub mod document;
 pub mod editor;
